Reconnect to last connected device
When the app is not connected to a device and a new device is discovered, the
app will automatically connect to the new device if the name of the device is
the same as the name of the last device that the app has been connected to.

Drop the "default device" concept in favor of simply keeping track of the last
connected device. This should improve the user experience as it is no longer
necessary to manually configure a default device.